refactor(statusbar): 抽通用 StatusChip + chip 按归属下沉到 features
StatusBar 内部杂乱、各 chip 自带逻辑、无公共 chip 组件，重构：
- 新增 common/StatusChip 通用 chip 壳（span/button + ok/err 色调 + 专属类名）
- 各业务 chip 按归属下沉到所属 feature 的 statusbar/：
  · settings: LlmChip · UserChip
  · chat: PrAgentActiveChip(+QueuePopover) · AutopilotChip
  · pr: LastSyncChip · RepoSyncChip · PrsCountChip
- pr-agent 运行时 chip 与更新 chip（app 运行时级）留 layout
- StatusBar 退化为薄壳（610→146 行），只做组合与布局
- 时长 / 相对时间格式化统一抽到 utils/time（formatElapsed 带 compact
  选项、formatRelative），消除与 chat 的重复实现（DRY）

无逻辑 / 界面变更。
